net: networkproxy: introduce network proxy framework
Network proxy maintains network presence while Host is in low-power
mode. Host is woke-up when there is print job or other packet of
interest not handled by Network Proxy agent.

This technology is described in ECMA-393 proxZzzy. It requires two
devices to be registered to the Network Proxy framework: IPC Device
and Network Device.

Signed-off-by: Ong Boon Leong <boon.leong.ong@intel.com>
Signed-off-by: Lay Kuan Loon <kuan.loon.lay@intel.com>
Signed-off-by: Song Yoong Siang <yoong.siang.song@intel.com>
Signed-off-by: Noor Azura Ahmad Tarmizi <noor.azura.ahmad.tarmizi@intel.com>